+++ stationery/dashboard/templates/setting/item/index.html
@@ -36,7 +36,7 @@
                 <td>{{ item.item_id }}</td>
                 <td>{{ item.item_name }}</td>
                 <td>{{ item.item_desc }}</td>
-                <td>{{ item.cate_id_id }}</td>
+                <td>{{ item.cate_id }}</td>
                 <td>
                   {% if item.item_img %}
                       <img src="/media/{{item.item_img }}" style="width: 50px; height: 50px;" alt="">
@@ -54,7 +54,7 @@
                     >Edit</a
                   >
                   <a
-                    onclick="delete_item('{{ item.id }}')"
+                    onclick="delete_item('{{ item.item_id }}')"
                     class="btn btn-danger "
                   >
                     Delete
@@ -89,7 +89,7 @@
       confirmButtonText: "Yes, delete it!",
     }).then((result) => {
       if (result.isConfirmed) {
-        window.location.href = "supplier/delete/" + id + "/";
+        window.location.href = "item/delete/" + id + "/";
       }
     });
   }
