--- stationery/dashboard/views/user_views.py
@@ -1,3 +1,0 @@
-from django.shortcuts import render ,redirect
-from .models import *
-import sweetify # type: ignore
